Add TableSlice react component
