+++ libs/cms/src/lib/environments/environment.ts
@@ -61,6 +61,7 @@ export default {
     'formField',
     'graphCard',
     'powerBiSlice',
+    'tableSlice',
   ],
   contentful: {
     space: process.env.CONTENTFUL_SPACE || '8k0h54kbe6bj',
